requirements-dev.txt
-r requirements.txt

coverage==4.5.1
django-nose==1.4.6
mock==2.0.0
flake8==3.4.1
pytest>=3.9.1
pytest-cov>=2.6.0
pytest-django>=3.4.3
